move fee conversion to separate mod
